Add progressive darkening on multiple hover
The function changeAlphaValue accept the step increase value to increase the
transparency of the grid elements when hovered.
A regular expression is used to match and store the values from the current element background color (i.e. digits correspondig to RGBa).
The alpha value is then updated and a new background color value is assigned using template literals.

The eventListener is now attached to the grid elements.
Attaching the hover eventListener to the parent element (container), lead to a bug of progressive darkening of the parent element too.
The bug could be fized digging deeper in the event propagation logic and corresponding methods.